fix(site): break agent detail lint cycle
